rag-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
azure-search-documents==11.4.0
azure-identity==1.15.0
openai==1.12.0
httpx==0.27.0
python-dotenv==1.0.0
python-multipart==0.0.6